+++ src/main/java/com/biopark/grupo2/model/Formulario.java
@@ -17,7 +17,7 @@ public class Formulario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_formulario")
-    private int id_formulario;
+    private long id_formulario;
 
     @Column(name = "titulo")
     private String titulo;
